Improved comments.

+++ TheatreProject/TheatreProject/Helpers/Paginator.cs
@@ -4,6 +4,7 @@ using System.Linq;
 
 namespace TheatreProject.Helpers
 {
+    // Handles splitting collections into pages
     public class Paginator<T> : IPaginator
     {
         private int currentPage;
@@ -43,7 +44,7 @@ namespace TheatreProject.Helpers
         public Paginator(IQueryable<T> items, int currentPage, int postsPerPage)
         {
             this.currentPage = currentPage;
-            this.totalPages = (int)Math.Ceiling((double)items.Count() / (double)postsPerPage) - 1;
+            this.totalPages = (int)Math.Ceiling((double)items.Count() / postsPerPage) - 1;
             int itemsToSkip = this.currentPage * postsPerPage;
             this.items = items.Skip(itemsToSkip).Take(postsPerPage).ToList();
         }
